Updates project API version and default value (#1285)
* Updates default API version to 64.0

This ensures that new projects and users are leveraging the most recent features and improvements offered by the Salesforce platform.

* Updates project API version during backup

Ensures the project's API version in sfdx-project.json and manifest files is up-to-date with the latest version defined in the environment or a fallback.

This prevents issues when deploying metadata to orgs with newer API versions and avoids manual updates. The update only happens if the current API version is lower than the target API version.

* changelog