Adds Ubuntu-based Docker images documentation
Adds documentation for Ubuntu-based Docker images to the README, providing instructions and links for both Docker Hub and GitHub Packages. This enhances the CI/CD options available to users by providing images that are compatible with GitHub, Azure, and Bitbucket.
Also adds a reference to Ubuntu images to the changelog.