Enhances WebSocket reliability and adds Ubuntu Docker docs (#1369)
* Adds Ubuntu-based Docker images documentation

Adds documentation for Ubuntu-based Docker images to the README, providing instructions and links for both Docker Hub and GitHub Packages. This enhances the CI/CD options available to users by providing images that are compatible with GitHub, Azure, and Bitbucket.
Also adds a reference to Ubuntu images to the changelog.

* Wait for websocket server to be initialized

* changelog

* Improves WebSocket initialization reliability

Increases the number of retries and reduces the retry interval
to improve the reliability of WebSocket initialization. This
addresses potential issues where the WebSocket may not be
fully initialized before it is used, especially in environments
with slower network connections or resource constraints.