Adds data workspace support to sandbox refresh
Enables saving and restoring data during sandbox refresh using SFDMU projects.

This change introduces a 'dataWorkspaces' property in the sfdx-hardis.json schema, allowing users to specify a list of data workspaces to be exported before a sandbox refresh and imported after the refresh is complete.

The before-refresh command now exports data from the specified workspaces, and the after-refresh command imports the data into the refreshed sandbox.