Clarifies JSON schema descriptions (#1394)
* Sets initPermissionSets to array of strings

Updates the initPermissionSets configuration property to accept an array of strings.

This change simplifies the configuration and ensures that only string values representing permission set names or permission set group names are accepted.

* Clarifies schema descriptions with examples

Adds examples to the descriptions of `initPermissionSets` and `scratchOrgInitApexScripts` properties in the schema.

This improves clarity and provides guidance for users configuring these options.