improve file exporter and importer features (#1406)
* Send progress events to WebSocket server for better UI in VsCode

* Improve file export summary log

* Better progress

* Improve log

* Adds CSV logging for file export operations

Implements detailed CSV logging to track the status, folder, filename,
extension, and size of exported files, including error details when
applicable.

This provides a comprehensive report of the file export process,
improving transparency and facilitating debugging. It also includes
a WebSocket message to notify users about the log file.

* changelog

* import (not validated)

* log

* polishing

* Updates progress message parameters

Updates progress messages to use the number of processed files and the total number of files instead of the command object.
This change ensures the correct parameters are used for progress tracking.

* changelog

* [Mega-Linter] Apply linters fixes

* changelog

* Fixes typo in XLSX file extension references

Corrects a typographical error by changing all instances of "XSLX" to "XLSX" in log messages and file names related to the generated Excel report files.

* Enhances work save command

Modifies the work save command to display the manual actions file as an action link in the UI, improving user experience by providing direct access to the file.

Also updates the changelog to reflect this change and a change to hardis org file commands.

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>