Improves Azure DevOps image attachment handling (#1451)
* Improves Azure DevOps image attachment handling

Addresses potential issues with a large number of attached images to a work item in Azure DevOps.
If the number of attached images reaches a threshold, it renames the current work item and creates a new one to avoid exceeding limitations.

Also, it adds a try-catch block when uploading images to Azure, in order to properly log any potential error.

* [Mega-Linter] Apply linters fixes

* changelog

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>