Fixes default org selection prompt (#1610)
* Fixes default org selection prompt

Addresses an issue where the response to the default org prompt was being ignored, causing the org to always be set as default, regardless of the user's choice.

This change ensures that the setDefault flag is correctly evaluated based on the user's response.

Fixes https://github.com/hardisgroupcom/sfdx-hardis/issues/1601

* fixes